Move the contents of load.go into cache.go
